use channel to communicate exit reasons
In order to streamline handling of the various exit reasons,
split out process lifecycle and timer.

Also use channels to tell WHY we need to exit now.
In case of timeout killing, we now collect the exit code of the dead
process and cleanly delete the lock instead of littering the system.