Use multiple channels in core loop
In order to react on multiple concurrent events in the core loop,
we need to use more channels and switch them on/off as needed.

We run as long as we have a process to nurse (watched at errc) and react
on other events while waiting for this process.

This make it easier to implement the following features:
 * soft kills (as just another timer)
 * retry of soft kills (timer + counter)
 * signal handling (just another channel)
 * retry of the command as new channel + new logstream + new new
   processLife)