More flexible monitoring sender/receiver
In order to support sending events as multiple entities per host
and in order to send to specific monitoring servers for specific
entities/events.

There is neither a default for send-as not for send-to. If you want it
static, then don't use these options and hard-code the senders and/or
receivers into the config.

The options `--send-as` and `--send-to` are free form strings.
They can even be URLs and contain spaces if your monitoring solution
works this way.